refactor: move verification code into gw_generator::verification
