+++ crates/generator/src/constants.rs
@@ -1,7 +1,7 @@
-// TODO ensure this value
-pub const MIN_WITHDRAWAL_CAPACITY: u64 = 1000_00000000;
-// TODO ensure this value
-pub const MIN_DEPOSIT_CAPACITY: u64 = 1000_00000000;
+/// MAX tx size 50 KB
+pub const MAX_TX_SIZE: usize = 50_000;
+/// MAX withdrawal size 50 KB
+pub const MAX_WITHDRAWAL_SIZE: usize = 50_000;
 // 25 KB
 pub const MAX_WRITE_DATA_BYTES_LIMIT: usize = 25_000;
 // 2MB
